moving some types from store to lib-types to prepare account summary storage
